bind9.preinst: cope with ancient conffile named.conf.options
In squeeze and earlier /etc/bind/named.conf.options was a conffile.
On upgrades from such a version dpkg will remember it as an obsolete
conffile with an outdated md5sum. Therefore we must not move it aside in
this case (which dpkg would take as a deletion by the local admin),
instead we edit it in place to match the to-be-installed version in
order to avoid prompting and make dpkg update the md5sum.

Closes: #905177